requirements.txt
asgiref==3.8.1
asn1crypto==1.5.1
astroid==3.2.4
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
autopep8==2.3.1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
crispy-bootstrap4==2024.1
cryptography==43.0.0
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
Django==5.1
django-allauth==64.1.0
django-appconf==1.0.6
django-countries==7.6.1
django-crispy-forms==2.3
django-tinymce-lite==0.0.2
filelock==3.15.4
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-recaptcha-enterprise==1.22.0
googleapis-common-protos==1.65.0
googlesearch-python==1.2.5
grpcio==1.66.1
grpcio-status==1.66.1
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
isort==5.13.2
jsmin==3.0.1
lazy-object-proxy==1.10.0
mccabe==0.7.0
oauthlib==3.2.2
pep8==1.7.1
pillow==10.4.0
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==5.28.0
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.1
pycparser==2.22
PyHamcrest==2.1.0
PyJWT==2.9.0
pylint==3.2.6
pyparsing==3.1.4
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==73.0.1
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
stripe==10.8.0
tomlkit==0.13.2
txaio==23.1.1
typed_ast==1.5.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.3
virtualenvwrapper-win==1.2.7
wrapt==1.16.0
zope.interface==7.0.1
